chore(gpu): plug in signed gpu tests in the hl api
